refactored Button component to use new structure
