+++ src/components/button/index.ts
@@ -1,6 +1,6 @@
-import Button from './Button';
+import ButtonRoot from './Button';
 
 export * from './Button.types';
 export * from './Button.props';
 
-export default Button;
+export default ButtonRoot;
